moved user tests to customer and added correct user tests
