+++ src/test/elements/autotaskcrm/assets/customers.json
@@ -1,0 +1,11 @@
+{
+  "clientPortalActive": true,
+  "contactID": 4,
+  "dateFormat": 1,
+  "id": "0",
+  "numberFormat": 22,
+  "password": "yourmom",
+  "securityLevel": 5,
+  "timeFormat": 19,
+  "userName": "mrchurros@cloud-elements.com"
+}
